LPD-46267 Move constants to API module instead of accessing through a helper
