+++ modules/apps/staging/staging-api/src/main/java/com/liferay/staging/StagingGroupHelper.java
@@ -33,8 +33,6 @@ public interface StagingGroupHelper {
 
 	public Group fetchRemoteLiveGroup(long groupId);
 
-	public String getCompanyGroupFriendlyURL();
-
 	public Group getStagedPortletGroup(Group group, String portletId);
 
 	public long getStagedPortletGroupId(long groupId, String portletId);
