+++ modules/apps/staging/staging-api/src/main/java/com/liferay/staging/constants/CompanyGroupConstants.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.staging.internal.constants;
+package com.liferay.staging.constants;
 
 /**
  * @author Alejandro Tardín
